requirements.txt
aiogram==3.20.0.post0
boto3==1.39.0
cachetools==6.1.0
celery==5.5.3
celery-types==0.23.0
colorlog==6.9.0
dj_database_url==2.3.0
Django==5.1.7
django-axes[ipware]==8.0.0
django-cors-headers==4.7.0
django-silk==5.4.0
django-split-settings==1.3.2
django-storages==1.14.5
django_redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
gunicorn==23.0.0
psycopg[c]==3.2.9
pyjwt==2.10.1
python-dotenv==1.1.1
pytz==2025.2
redis==6.2.0
sentry-sdk==2.32.0
pyngrok>=5.0.5
